+++ cedar-policy-core/src/parser/text_to_cst.rs
@@ -1012,7 +1012,7 @@ mod tests {
     }
 
     #[test]
-    fn policy_annotations() {
+    fn policy_annotations_ok() {
         let policies = assert_parse_succeeds(
             parse_policies,
             r#"
@@ -1032,7 +1032,25 @@ mod tests {
                 .len(),
             4
         );
+    }
 
+    #[test]
+    fn policy_annotations_no_value_ok() {
+        let policy = assert_parse_succeeds(
+            parse_policy,
+            r#"@foo permit (principal, action, resource);"#,
+        );
+        let annotation = policy.annotations.first().unwrap().as_inner().unwrap();
+        assert_eq!(annotation.value, None);
+        assert_eq!(
+            annotation.key.as_inner().unwrap().to_string(),
+            "foo".to_string()
+        );
+        assert_eq!(policy.annotations.len(), 1);
+    }
+
+    #[test]
+    fn policy_annotations_bad_id() {
         let src = r#"
             @bad-annotation("bad") permit (principal, action, resource);
         "#;
@@ -1041,10 +1059,42 @@ mod tests {
             src,
             &errs,
             &ExpectedErrorMessageBuilder::error("unexpected token `-`")
-                .exactly_one_underline_with_label("-", "expected `(`")
+                .exactly_one_underline_with_label("-", "expected `(`, `@`, or identifier")
                 .build(),
         );
 
+        let src = r#"
+            @hi mom("this should be invalid")
+            permit(principal, action, resource);
+        "#;
+        let errs = assert_parse_fails(parse_policies, src);
+        expect_exactly_one_error(
+            src,
+            &errs,
+            &ExpectedErrorMessageBuilder::error("unexpected token `\"this should be invalid\"`")
+                .exactly_one_underline_with_label(
+                    "\"this should be invalid\"",
+                    "expected `)` or identifier",
+                )
+                .build(),
+        );
+
+        let src = r#"
+            @hi+mom("this should be invalid")
+            permit(principal, action, resource);
+        "#;
+        let errs = assert_parse_fails(parse_policies, src);
+        expect_exactly_one_error(
+            src,
+            &errs,
+            &ExpectedErrorMessageBuilder::error("unexpected token `+`")
+                .exactly_one_underline_with_label("+", "expected `(`, `@`, or identifier")
+                .build(),
+        );
+    }
+
+    #[test]
+    fn policy_annotations_bad_val() {
         let src = r#"
             @bad_annotation("bad","annotation") permit (principal, action, resource);
         "#;
@@ -1054,6 +1104,18 @@ mod tests {
             &errs,
             &ExpectedErrorMessageBuilder::error("unexpected token `,`")
                 .exactly_one_underline_with_label(",", "expected `)`")
+                .build(),
+        );
+
+        let src = r#"
+            @bad_annotation() permit (principal, action, resource);
+        "#;
+        let errs = assert_parse_fails(parse_policies, src);
+        expect_exactly_one_error(
+            src,
+            &errs,
+            &ExpectedErrorMessageBuilder::error("unexpected token `)`")
+                .exactly_one_underline_with_label(")", "expected string literal")
                 .build(),
         );
 
@@ -1068,7 +1130,10 @@ mod tests {
                 .exactly_one_underline_with_label("bad_annotation", "expected string literal")
                 .build(),
         );
+    }
 
+    #[test]
+    fn policy_annotation_bad_position() {
         let src = r#"
             permit (@comment("your name here") principal, action, resource);
         "#;
@@ -1078,32 +1143,6 @@ mod tests {
             &errs,
             &ExpectedErrorMessageBuilder::error("unexpected token `@`")
                 .exactly_one_underline_with_label("@", "expected `)` or identifier")
-                .build(),
-        );
-
-        let src = r#"
-            @hi mom("this should be invalid")
-            permit(principal, action, resource);
-        "#;
-        let errs = assert_parse_fails(parse_policies, src);
-        expect_exactly_one_error(
-            src,
-            &errs,
-            &ExpectedErrorMessageBuilder::error("unexpected token `mom`")
-                .exactly_one_underline_with_label("mom", "expected `(`")
-                .build(),
-        );
-
-        let src = r#"
-            @hi+mom("this should be invalid")
-            permit(principal, action, resource);
-        "#;
-        let errs = assert_parse_fails(parse_policies, src);
-        expect_exactly_one_error(
-            src,
-            &errs,
-            &ExpectedErrorMessageBuilder::error("unexpected token `+`")
-                .exactly_one_underline_with_label("+", "expected `(`")
                 .build(),
         );
     }